Fix custom converter sample
Fixes #93
