+++ custom_converter/starter.py
@@ -1,28 +1,22 @@
 import asyncio
-import dataclasses
 
-import temporalio.converter
 from temporalio.client import Client
 
-from custom_converter.worker import (
+from custom_converter.shared import (
     GreetingInput,
     GreetingOutput,
-    GreetingPayloadConverter,
-    GreetingWorkflow,
+    greeting_data_converter,
 )
+from custom_converter.workflow import GreetingWorkflow
 
 
 async def main():
     # Connect client
     client = await Client.connect(
         "localhost:7233",
-        # Use the default data converter, but change the payload converter.
         # Without this we get:
         #   TypeError: Object of type GreetingInput is not JSON serializable
-        data_converter=dataclasses.replace(
-            temporalio.converter.default(),
-            payload_converter_class=GreetingPayloadConverter,
-        ),
+        data_converter=greeting_data_converter,
     )
 
     # Run workflow
